fix: Move methods to match grids to separate module.
