+++ src/porepy/utils/grid_refinement.py
@@ -1,4 +1,3 @@
-import os
 import logging
 import gmsh
 from typing import (  # noqa
